refactor(doctor): extract uv/home/virtualenv/symlink checks + registration factory
