+++ src/core/doctor/mod.rs
@@ -3,15 +3,11 @@
 //! This module provides functionality to diagnose the scuv installation
 //! and report any issues with suggested fixes.
 
+mod checks;
 mod engine;
 mod types;
 
-use std::process::Command;
-
-use crate::core::metadata::Metadata;
 use crate::paths;
-use crate::uv::UvClient;
-use crate::uv::version as uv_version;
 
 pub use engine::Doctor;
 pub use types::{Check, CheckResult, CheckStatus};
@@ -19,497 +15,6 @@ pub use types::{Check, CheckResult, CheckStatus};
 // ============================================================================
 // Individual Checks
 // ============================================================================
-
-/// Check for uv installation.
-struct UvCheck;
-
-impl UvCheck {
-    /// Platform-appropriate install or upgrade command for uv.
-    fn install_hint() -> &'static str {
-        if cfg!(target_os = "macos") {
-            "brew install uv  OR  curl -LsSf https://astral.sh/uv/install.sh | sh"
-        } else if cfg!(target_os = "windows") {
-            "powershell -ExecutionPolicy ByPass -c \"irm https://astral.sh/uv/install.ps1 | iex\""
-        } else {
-            "curl -LsSf https://astral.sh/uv/install.sh | sh"
-        }
-    }
-}
-
-impl Check for UvCheck {
-    fn id(&self) -> &'static str {
-        "uv"
-    }
-
-    fn name(&self) -> &'static str {
-        "uv installation"
-    }
-
-    fn run(&self) -> Vec<CheckResult> {
-        match Command::new("uv").arg("--version").output() {
-            Ok(output) if output.status.success() => {
-                let raw = String::from_utf8_lossy(&output.stdout);
-                let raw = raw.trim();
-
-                // Enforce the minimum supported uv version when parseable.
-                // Unparseable output (custom build, unknown format) is treated
-                // as a soft pass so we don't block users on a banner change.
-                if let Some(version) = uv_version::parse(raw) {
-                    if !uv_version::meets_minimum(version) {
-                        return vec![
-                            CheckResult::error(
-                                self.id(),
-                                self.name(),
-                                format!(
-                                    "uv {} is older than the supported minimum ({})",
-                                    uv_version::format_version(version),
-                                    uv_version::format_version(uv_version::MIN_VERSION),
-                                ),
-                            )
-                            .with_details(raw.to_string())
-                            .with_suggestion(format!("Upgrade uv: {}", Self::install_hint())),
-                        ];
-                    }
-                }
-
-                vec![CheckResult::ok(self.id(), self.name()).with_details(raw.to_string())]
-            }
-            _ => {
-                vec![
-                    CheckResult::error(self.id(), self.name(), "uv not found in PATH")
-                        .with_details("scuv requires uv to manage Python environments")
-                        .with_suggestion(format!("Install uv: {}", Self::install_hint())),
-                ]
-            }
-        }
-    }
-}
-
-/// Check for SCUV_HOME directory.
-struct HomeCheck;
-
-impl Check for HomeCheck {
-    fn id(&self) -> &'static str {
-        "home"
-    }
-
-    fn name(&self) -> &'static str {
-        "SCUV_HOME directory"
-    }
-
-    fn run(&self) -> Vec<CheckResult> {
-        match paths::scoop_home() {
-            Ok(path) if path.exists() => {
-                // Check write permission
-                match path.metadata() {
-                    Ok(meta) if !meta.permissions().readonly() => {
-                        vec![
-                            CheckResult::ok(self.id(), self.name())
-                                .with_details(format!("{}", path.display())),
-                        ]
-                    }
-                    _ => {
-                        vec![
-                            CheckResult::error(self.id(), self.name(), "directory not writable")
-                                .with_suggestion(format!("chmod 755 {}", path.display())),
-                        ]
-                    }
-                }
-            }
-            Ok(path) => {
-                vec![
-                    CheckResult::error(self.id(), self.name(), "directory not found")
-                        .with_suggestion(format!("mkdir -p {}", path.display())),
-                ]
-            }
-            Err(_) => {
-                vec![
-                    CheckResult::error(
-                        self.id(),
-                        self.name(),
-                        "could not determine home directory",
-                    )
-                    .with_suggestion("Set SCUV_HOME environment variable"),
-                ]
-            }
-        }
-    }
-
-    fn fix(&self, result: &CheckResult, output: &crate::output::Output) -> Option<CheckResult> {
-        // Only fix "directory not found" errors
-        if let CheckStatus::Error(msg) = &result.status {
-            if msg.contains("not found") {
-                // Create the directory
-                if let Ok(home) = paths::scoop_home() {
-                    output.info(&format!("Creating {}...", home.display()));
-
-                    match std::fs::create_dir_all(&home) {
-                        Ok(_) => {
-                            // Also create virtualenvs subdirectory
-                            let _ = std::fs::create_dir_all(home.join("virtualenvs"));
-
-                            return Some(
-                                CheckResult::ok("home", "SCUV_HOME directory")
-                                    .with_details(format!("created {}", home.display())),
-                            );
-                        }
-                        Err(e) => {
-                            return Some(
-                                CheckResult::error(
-                                    "home",
-                                    "SCUV_HOME directory",
-                                    format!("failed to create: {}", e),
-                                )
-                                .with_suggestion("Check permissions"),
-                            );
-                        }
-                    }
-                }
-            }
-        }
-        None
-    }
-}
-
-/// Check for virtualenv integrity.
-struct VirtualenvCheck;
-
-impl Check for VirtualenvCheck {
-    fn id(&self) -> &'static str {
-        "venv"
-    }
-
-    fn name(&self) -> &'static str {
-        "virtual environments"
-    }
-
-    fn run(&self) -> Vec<CheckResult> {
-        let venvs_dir = match paths::virtualenvs_dir() {
-            Ok(dir) => dir,
-            Err(_) => {
-                return vec![CheckResult::error(
-                    self.id(),
-                    self.name(),
-                    "virtualenvs directory not found",
-                )];
-            }
-        };
-
-        if !venvs_dir.exists() {
-            return vec![
-                CheckResult::ok(self.id(), self.name()).with_details("no environments yet"),
-            ];
-        }
-
-        let mut results = Vec::new();
-        let mut healthy = 0;
-        let mut broken_names = Vec::new();
-
-        if let Ok(entries) = std::fs::read_dir(&venvs_dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.is_dir() {
-                    let name = path
-                        .file_name()
-                        .and_then(|n| n.to_str())
-                        .unwrap_or("unknown")
-                        .to_string();
-
-                    let python_path = crate::paths::virtualenv_python_exe(&path);
-                    let pyvenv_cfg = path.join("pyvenv.cfg");
-
-                    if python_path.exists() && pyvenv_cfg.exists() {
-                        healthy += 1;
-                    } else {
-                        broken_names.push(name);
-                    }
-                }
-            }
-        }
-
-        // Report broken environments
-        for name in &broken_names {
-            results.push(
-                CheckResult::error(
-                    "venv",
-                    "broken virtualenv",
-                    format!("'{}' is corrupted", name),
-                )
-                .with_suggestion(format!(
-                    "scuv remove {} && scuv create {} <python-version>",
-                    name, name
-                )),
-            );
-        }
-
-        // Summary
-        if broken_names.is_empty() {
-            if healthy > 0 {
-                results.push(
-                    CheckResult::ok(self.id(), self.name())
-                        .with_details(format!("{} environments, all healthy", healthy)),
-                );
-            } else {
-                results.push(
-                    CheckResult::ok(self.id(), self.name()).with_details("no environments yet"),
-                );
-            }
-        }
-
-        results
-    }
-}
-
-/// Check for symbolic link validity.
-struct SymlinkCheck;
-
-impl Check for SymlinkCheck {
-    fn id(&self) -> &'static str {
-        "symlink"
-    }
-
-    fn name(&self) -> &'static str {
-        "symbolic links"
-    }
-
-    fn run(&self) -> Vec<CheckResult> {
-        let venvs_dir = match paths::virtualenvs_dir() {
-            Ok(dir) if dir.exists() => dir,
-            _ => return vec![],
-        };
-
-        let mut results = Vec::new();
-        let mut valid = 0;
-        let mut broken_names = Vec::new();
-
-        if let Ok(entries) = std::fs::read_dir(&venvs_dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.is_dir() {
-                    let python_path = crate::paths::virtualenv_python_exe(&path);
-
-                    if python_path.is_symlink() {
-                        match std::fs::read_link(&python_path) {
-                            Ok(target) if target.exists() => {
-                                valid += 1;
-                            }
-                            Ok(_) => {
-                                let name = path
-                                    .file_name()
-                                    .and_then(|n| n.to_str())
-                                    .unwrap_or("unknown")
-                                    .to_string();
-                                broken_names.push(name);
-                            }
-                            Err(_) => {
-                                // Not a symlink or error reading
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Report broken symlinks
-        for name in &broken_names {
-            results.push(
-                CheckResult::error(
-                    "symlink",
-                    "broken symlink",
-                    format!("Python symlink in '{}' is broken", name),
-                )
-                .with_suggestion(format!(
-                    "scuv remove {} && scuv create {} <python-version>",
-                    name, name
-                )),
-            );
-        }
-
-        // Summary
-        if broken_names.is_empty() && valid > 0 {
-            results.push(
-                CheckResult::ok(self.id(), self.name())
-                    .with_details(format!("{} symlinks valid", valid)),
-            );
-        }
-
-        results
-    }
-
-    fn fix(&self, result: &CheckResult, output: &crate::output::Output) -> Option<CheckResult> {
-        // Extract environment name from error message: "Python symlink in 'name' is broken"
-        let venv_name = if let CheckStatus::Error(msg) = &result.status {
-            // Parse: "Python symlink in 'name' is broken"
-            msg.split('\'').nth(1).map(|s| s.to_string())
-        } else {
-            None
-        }?;
-
-        output.info(&format!("Attempting to fix symlink for '{}'...", venv_name));
-
-        // Get virtualenv path
-        let venvs_dir = paths::virtualenvs_dir().ok()?;
-        let venv_path = venvs_dir.join(&venv_name);
-
-        if !venv_path.exists() {
-            return Some(
-                CheckResult::error(
-                    "symlink",
-                    "broken symlink",
-                    format!("environment '{}' not found", venv_name),
-                )
-                .with_suggestion(format!("scuv create {} <python-version>", venv_name)),
-            );
-        }
-
-        // Read metadata to get Python version
-        let metadata_path = venv_path.join(Metadata::FILE_NAME);
-        let python_version = if metadata_path.exists() {
-            match std::fs::read_to_string(&metadata_path) {
-                Ok(content) => match serde_json::from_str::<Metadata>(&content) {
-                    Ok(meta) => Some(meta.python_version),
-                    Err(_) => None,
-                },
-                Err(_) => None,
-            }
-        } else {
-            // Try to extract from pyvenv.cfg as fallback
-            let pyvenv_cfg = venv_path.join("pyvenv.cfg");
-            if pyvenv_cfg.exists() {
-                std::fs::read_to_string(&pyvenv_cfg)
-                    .ok()
-                    .and_then(|content| {
-                        for line in content.lines() {
-                            if line.starts_with("version") {
-                                return line.split('=').nth(1).map(|v| v.trim().to_string());
-                            }
-                        }
-                        None
-                    })
-            } else {
-                None
-            }
-        };
-
-        let python_version = match python_version {
-            Some(v) => v,
-            None => {
-                return Some(
-                    CheckResult::error(
-                        "symlink",
-                        "broken symlink",
-                        format!("could not determine Python version for '{}'", venv_name),
-                    )
-                    .with_suggestion(format!(
-                        "scuv remove {} && scuv create {} <python-version>",
-                        venv_name, venv_name
-                    )),
-                );
-            }
-        };
-
-        output.info(&format!("Found Python version: {}", python_version));
-
-        // Find Python binary using uv
-        let uv = match UvClient::new() {
-            Ok(uv) => uv,
-            Err(_) => {
-                return Some(
-                    CheckResult::error("symlink", "broken symlink", "uv not available")
-                        .with_suggestion("Install uv first"),
-                );
-            }
-        };
-
-        let python_path = match uv.find_python(&python_version) {
-            Ok(Some(info)) => match info.path {
-                Some(path) => path,
-                None => {
-                    return Some(
-                        CheckResult::error(
-                            "symlink",
-                            "broken symlink",
-                            format!("Python {} path not found", python_version),
-                        )
-                        .with_suggestion(format!("scuv install {}", python_version)),
-                    );
-                }
-            },
-            Ok(None) => {
-                return Some(
-                    CheckResult::error(
-                        "symlink",
-                        "broken symlink",
-                        format!("Python {} not installed", python_version),
-                    )
-                    .with_suggestion(format!("scuv install {}", python_version)),
-                );
-            }
-            Err(_) => {
-                return Some(
-                    CheckResult::error(
-                        "symlink",
-                        "broken symlink",
-                        "failed to find Python installation",
-                    )
-                    .with_suggestion(format!("scuv install {}", python_version)),
-                );
-            }
-        };
-
-        // Recreate symlink
-        let symlink_path = crate::paths::virtualenv_python_exe(&venv_path);
-
-        // Remove old symlink if exists
-        if symlink_path.exists() || symlink_path.is_symlink() {
-            if let Err(e) = std::fs::remove_file(&symlink_path) {
-                return Some(
-                    CheckResult::error(
-                        "symlink",
-                        "broken symlink",
-                        format!("failed to remove old symlink: {}", e),
-                    )
-                    .with_suggestion("Check file permissions"),
-                );
-            }
-        }
-
-        // Create new symlink
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::symlink;
-            if let Err(e) = symlink(&python_path, &symlink_path) {
-                return Some(
-                    CheckResult::error(
-                        "symlink",
-                        "broken symlink",
-                        format!("failed to create symlink: {}", e),
-                    )
-                    .with_suggestion("Check file permissions"),
-                );
-            }
-        }
-
-        #[cfg(not(unix))]
-        {
-            return Some(
-                CheckResult::warn(
-                    "symlink",
-                    "broken symlink",
-                    "symlink fix not supported on this platform",
-                )
-                .with_suggestion("Manually recreate the symlink"),
-            );
-        }
-
-        output.success(&format!("Fixed symlink for '{}'", venv_name));
-
-        Some(
-            CheckResult::ok("symlink", "broken symlink")
-                .with_details(format!("fixed symlink for '{}'", venv_name)),
-        )
-    }
-}
 
 /// Check for shell configuration (scuv init).
 struct ShellCheck;
@@ -934,86 +439,8 @@ mod tests {
         );
     }
 
-    // ==========================================================================
-    // VirtualenvCheck / SymlinkCheck / fix_symlink coverage
-    //
-    // These tests pin the run() / fix_symlink() implementations against
-    // the cargo-mutants `replace -> vec![]` / `replace -> None`
-    // mutations. Without them, deleting the entire function body would
-    // still pass the suite, masking real regressions.
-    // ==========================================================================
-
     use crate::test_utils::with_temp_scoop_home;
     use serial_test::serial;
-
-    #[test]
-    #[serial]
-    fn virtualenv_check_emits_broken_for_corrupted_env() {
-        with_temp_scoop_home(|temp| {
-            // Build a deliberately broken venv: directory exists but
-            // pyvenv.cfg is missing and there's no python binary. The
-            // check must surface this as a broken-virtualenv error
-            // (not silently return an empty vec or an Ok summary).
-            let broken = temp.path().join("virtualenvs").join("brokenenv");
-            std::fs::create_dir_all(&broken).unwrap();
-
-            let results = VirtualenvCheck.run();
-            assert!(!results.is_empty(), "expected at least one CheckResult");
-            assert!(
-                results.iter().any(|r| r.is_error()
-                    && r.name.contains("broken")
-                    && matches!(&r.status, CheckStatus::Error(msg) if msg.contains("brokenenv"))),
-                "expected an error CheckResult naming the broken env, got {results:#?}"
-            );
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn virtualenv_check_returns_ok_for_healthy_env() {
-        with_temp_scoop_home(|temp| {
-            let env = temp.path().join("virtualenvs").join("healthy");
-            std::fs::create_dir_all(env.join("bin")).unwrap();
-            std::fs::write(env.join("bin").join("python"), "").unwrap();
-            std::fs::write(env.join("pyvenv.cfg"), "").unwrap();
-
-            let results = VirtualenvCheck.run();
-            assert!(!results.is_empty(), "expected a non-empty result set");
-            assert!(
-                results.iter().all(|r| r.is_ok()),
-                "all results should be Ok, got {results:#?}"
-            );
-        });
-    }
-
-    /// On Unix we can deterministically create a symlink whose target
-    /// doesn't exist, which is exactly the failure SymlinkCheck::run
-    /// surfaces. cfg-gated to Unix because the symlink primitive
-    /// differs on Windows (and the CI matrix that exercises mutants is
-    /// Linux only — same rationale as the existing
-    /// test_virtualenv_exists_with_broken_symlink in paths.rs).
-    #[cfg(unix)]
-    #[test]
-    #[serial]
-    fn symlink_check_emits_error_for_broken_python_symlink() {
-        with_temp_scoop_home(|temp| {
-            use std::os::unix::fs::symlink;
-
-            let env = temp.path().join("virtualenvs").join("brokenlink");
-            std::fs::create_dir_all(env.join("bin")).unwrap();
-            symlink("/nonexistent/python", env.join("bin").join("python")).unwrap();
-
-            let results = SymlinkCheck.run();
-            assert!(
-                !results.is_empty(),
-                "broken symlink env must produce at least one result"
-            );
-            assert!(
-                results.iter().any(|r| r.is_error()),
-                "expected at least one error result, got {results:#?}"
-            );
-        });
-    }
 
     // ==========================================================================
     // VersionCheck: `system` sentinel handling
@@ -1207,39 +634,6 @@ mod tests {
         fn drop(&mut self) {
             let _ = std::env::set_current_dir(&self.original);
         }
-    }
-
-    #[test]
-    #[serial]
-    fn fix_symlink_returns_some_for_parseable_error_name() {
-        with_temp_scoop_home(|temp| {
-            // fix_symlink parses the env name out of a "Python symlink
-            // in 'name' is broken" message. If the parse succeeds and
-            // the env doesn't exist, it returns Some(error suggesting
-            // scuv create). The cargo-mutants -> None replacement
-            // would silently drop that guidance.
-            let broken = temp.path().join("virtualenvs");
-            std::fs::create_dir_all(&broken).unwrap();
-
-            let probe = CheckResult::error(
-                "symlink",
-                "broken symlink",
-                "Python symlink in 'fix-target' is broken".to_string(),
-            );
-            let output = crate::output::Output::new(0, true, true, false);
-
-            let fixed = SymlinkCheck.fix(&probe, &output);
-            assert!(fixed.is_some(), "fix_symlink must return Some");
-            let r = fixed.unwrap();
-            assert!(r.is_error() || r.is_warning());
-            // Suggestion text should point the user at `scuv create`.
-            assert!(
-                r.suggestion
-                    .as_deref()
-                    .is_some_and(|s| s.contains("scuv create"))
-                    || matches!(&r.status, CheckStatus::Error(msg) if msg.contains("fix-target"))
-            );
-        });
     }
 
     // ==========================================================================
@@ -1623,40 +1017,6 @@ mod tests {
         assert!(results[0].is_ok(), "clean env must be Ok, got {results:#?}");
     }
 
-    #[test]
-    #[serial]
-    fn home_check_trait_dispatch_ok_on_existing_writable_home() {
-        with_temp_scoop_home(|temp| {
-            let check: &dyn Check = &HomeCheck;
-            assert_eq!(check.id(), "home");
-            assert_eq!(check.name(), "SCUV_HOME directory");
-            let results = check.run();
-            assert_eq!(results.len(), 1, "got {results:#?}");
-            assert!(results[0].is_ok(), "got {results:#?}");
-            let details = results[0].details.as_deref().unwrap_or_default();
-            assert!(
-                details.contains(temp.path().to_str().unwrap()),
-                "details should name the home path, got {details:?}"
-            );
-        });
-    }
-
-    #[test]
-    fn uv_check_trait_dispatch_always_reports_under_uv_id() {
-        // Environment-tolerant: passes whether or not a usable `uv` is on
-        // PATH — what it pins is that run() actually runs and every result
-        // carries this check's identity.
-        let check: &dyn Check = &UvCheck;
-        assert_eq!(check.id(), "uv");
-        assert_eq!(check.name(), "uv installation");
-        let results = check.run();
-        assert!(!results.is_empty(), "run() must produce results");
-        assert!(
-            results.iter().all(|r| r.id == "uv"),
-            "all results must carry the uv id, got {results:#?}"
-        );
-    }
-
     // ==========================================================================
     // ShellCheck zsh branch (the `shell_name == "zsh"` boundary): the
     // suggestion must name `scuv init zsh`, not bash, and read .zshrc.
@@ -1703,53 +1063,5 @@ mod tests {
         let results = ShellCheck.run();
         assert_eq!(results.len(), 1, "got {results:#?}");
         assert!(results[0].is_ok(), "got {results:#?}");
-    }
-
-    // ==========================================================================
-    // Doctor::fix_home — creates the missing home (+ virtualenvs) only for
-    // "not found" errors, and leaves anything else alone.
-    // ==========================================================================
-
-    #[test]
-    #[serial]
-    fn fix_home_creates_missing_home_and_virtualenvs_dir() {
-        let tmp = tempfile::tempdir().unwrap();
-        let missing_home = tmp.path().join("scuvhome");
-        let _g = crate::test_utils::env_guard(&[(
-            paths::SCUV_HOME_ENV,
-            Some(missing_home.to_str().unwrap()),
-        )]);
-
-        let output = crate::output::Output::new(0, true, true, false);
-        let broken = CheckResult::error("home", "SCUV_HOME directory", "directory not found");
-
-        let fixed = HomeCheck.fix(&broken, &output);
-        let fixed = fixed.expect("a 'not found' home error must be fixable");
-        assert!(fixed.is_ok(), "got {fixed:#?}");
-        assert!(missing_home.is_dir(), "home dir must be created");
-        assert!(
-            missing_home.join("virtualenvs").is_dir(),
-            "virtualenvs subdir must be created"
-        );
-    }
-
-    #[test]
-    #[serial]
-    fn fix_home_ignores_non_not_found_results() {
-        let tmp = tempfile::tempdir().unwrap();
-        let _g = crate::test_utils::env_guard(&[(
-            paths::SCUV_HOME_ENV,
-            Some(tmp.path().to_str().unwrap()),
-        )]);
-
-        let output = crate::output::Output::new(0, true, true, false);
-        let unrelated = CheckResult::error("home", "SCUV_HOME directory", "permission denied");
-        assert!(HomeCheck.fix(&unrelated, &output).is_none());
-
-        let warning = CheckResult::warn("home", "SCUV_HOME directory", "directory not found");
-        assert!(
-            HomeCheck.fix(&warning, &output).is_none(),
-            "only Error status is fixable, not Warning"
-        );
     }
 }
